% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructSe.R
\name{constructSe}
\alias{constructSe}
\title{Construct SummarizedExperiment object from a matrix}
\usage{
constructSe(dat, assay = "raw", name = NULL, colData = NULL, rowData = NULL)
}
\arguments{
\item{dat}{A p*n data frame or a matrix object where p is the number of
metabolites and n is the number of samples}

\item{assay}{A name of assay to be labelled}

\item{name}{Name of the data}

\item{colData}{A DataFrame object. This should include type of replicates,
sample name and etc.}

\item{rowData}{A DataFrame object. This should include details of the metabolites.}
}
\value{
A SummarizedExperiment object
}
\description{
Construct SummarizedExperiment object from a matrix
}
