% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign.R
\name{expDesign}
\alias{expDesign}
\title{Generate experiment design}
\usage{
expDesign(n = 1000, br = 5, ncol = 10, nrow = 10)
}
\arguments{
\item{n}{number of samples for the design}

\item{br}{number of biological samples for batch replicates}

\item{ncol}{number of columns per plate}

\item{nrow}{number of samples per plate}
}
\value{
A matrix of experiment design
}
\description{
Generate experiment design
}
