% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv.R
\name{getRUV}
\alias{getRUV}
\title{RUV function for metabolomics data}
\usage{
getRUV(dat, assay, k = 5, rep, negCtl = NULL, newAssay = NULL)
}
\arguments{
\item{dat}{A SummarizedExperiment object}

\item{k}{k for RUVIII}

\item{assay_name}{An assay to perform RUV}

\item{ctrl}{type of replication to utilise for design matrix}

\item{new_assay}{A name of the newY in the SummarizedExperiment object}

\item{replicate_mat}{A user defined design matrix for RUVIII}
}
\description{
RUV function for metabolomics data
}
