% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv.R
\name{hierarchy}
\alias{hierarchy}
\title{Hierarchical RUV}
\usage{
hierarchy(
  dat_list,
  assay,
  k = 5,
  rep,
  newAssay = NULL,
  balanced = FALSE,
  negCtl = NULL,
  hOrder = NULL
)
}
\arguments{
\item{dat_list}{list of SummarizedExperiment objects}

\item{assay}{Name of assay to merge}

\item{k}{\code{k} for RUV}

\item{rep}{NULL}

\item{newAssay}{NULL}

\item{balanced}{a logical value}

\item{negCtl}{Type of controls to use}

\item{hOrder}{Order of hierarchical normalisation}
}
\description{
Hierarchical RUV
}
