% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hruv.R
\name{hruv}
\alias{hruv}
\title{hRUV - intra- and hierarchical inter-batch normalisation.}
\usage{
hruv(
  dat_list,
  assay,
  intra = c("rlm", "loess", "rlmShort", "loessShort"),
  inter = c("balanced", "concatenate"),
  intra_k = 5,
  inter_k = intra_k,
  pCtlName,
  negCtl,
  intra_rep,
  inter_rep,
  hOrder = NULL,
  newAssay = NULL
)
}
\arguments{
\item{dat_list}{A list of SummarizedExperiment data where each object
represents a batch.}

\item{assay}{An assay name to measure the missingness of the signals.}

\item{intra}{An intra-batch normalisation method.
See Description for details.}

\item{inter}{An inter-batch normalisation method. Either "balanced" or
"concatenate" method for hierarchical approach.}

\item{intra_k}{An intra-batch RUV-III parameter k value.}

\item{inter_k}{An inter-batch RUV-III parameter k value.}

\item{pCtlName}{A name of the variable in the colData of each
SummarizedExperiment data in the \code{dat_list}. This variable in colData
should be a numeric ID of which samples to use for robust smoother fitting.}

\item{negCtl}{A vector of row IDs for selection of stable matbolites to be
used as negative control in RUV.}

\item{intra_rep}{A name of the variable in the colData of each
SummarizedExperiment data in the \code{dat_list}. This variable in colData
should be a logical vector of whic samples are a intra-batch replicate}

\item{inter_rep}{A name of the variable in the colData of each
SummarizedExperiment data in the \code{dat_list}. This variable in colData
should be a logical vector of which samples are a inter-batch replicate}

\item{hOrder}{A vector of batch names from names of list \code{dat_list}.}

\item{newAssay}{A name of the new assay for cleaned (preprocessed) data.}
}
\value{
A normalised data as SummarizedExperiment object.
}
\description{
Hirarchical approach to remove unwanted variation across
multiple batches of data.

For \code{intra} method, \code{rlm} or \code{loess} methods fits robust
linear model or loess respectively to samples selected from \code{pCtName}.
\code{rlmShort} and \code{loessShort} both performs robust smoother \code{rlm}
and \code{loess} respectively then performs RUV-III utilising the replicates
identified from \code{intra_rep}. Parameters \code{negCtl} and \code{intra_k}
will be used for RUV-III.
}
