% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraNorm.R
\name{intraNorm}
\alias{intraNorm}
\title{Intra-batch normalisation}
\usage{
intraNorm(
  dat_list,
  assay,
  method = c("rlm", "loess", "rlmShort", "loessShort"),
  pCtlName = NULL,
  rep = NULL,
  k,
  negCtl = NULL
)
}
\arguments{
\item{dat_list}{A list of SummarizedExperiment data where each object
represents a batch.}

\item{assay}{An assay name to measure the missingness of the signals.}

\item{method}{A method to use for removing signal drift and unwanted
variations. See descriptions for further details.}

\item{pCtlName}{A column variable name in colData of each
SummarizedExperiment object which selects samples to be used to fit the
robust smoother.}

\item{rep}{A column variable name in colData of each SummarizedExperiment
object which identifies intra-batch (short) sample replicates.}

\item{k}{A RUV-III paramater k value}

\item{negCtl}{A vector of row IDs for selection of stable matbolites to be
used as negative control in RUV.}
}
\value{
A list of intra-batch normalized SummarizedExperiment data.
}
\description{
Removes unwanted variation within batch.

\code{rlm} and \code{loess}

\code{rlm} or \code{loess} methods fits robust linear model or loess
respectively to samples selected from \code{pCtName}. \code{rlmShort} and
\code{loessShort} both performs robust smoother \code{rlm} and \code{loess}
respectively then performs RUV-III utilising the replicates identified from
\code{intra_rep}. Parameters \code{negCtl} and \code{intra_k} will be used
for RUV-III.
}
