% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRLE}
\alias{plotRLE}
\title{Relative Log Expression (RLE) plot}
\usage{
plotRLE(dat, assay, group = NULL)
}
\arguments{
\item{dat}{SummarizedExperiment object data}

\item{assay}{An assay to use in a SummarizedExperiment object}

\item{group}{A vector to assess group median.}
}
\value{
RLE plot
}
\description{
Relative Log Expression (RLE) plot
}
