% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRun}
\alias{plotRun}
\title{Metabolite run plot}
\usage{
plotRun(dat, assay, colour)
}
\arguments{
\item{dat}{SummarizedExperiment object data}

\item{assay}{An assay to use in a SummarizedExperiment object}

\item{colour}{A group vector to colour the points in the PCA plot}
}
\value{
A list of run plots
}
\description{
Metabolite run plot
}
