% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean each data}
\usage{
clean(
  dat_list,
  threshold = 0.5,
  method = c("intersect", "union"),
  assay = "raw",
  newAssay = assay
)
}
\arguments{
\item{dat_list}{A list of SummarizedExperiment data where each object
represents a batch.}

\item{threshold}{A threshold value between 0 and 1 of missingness to filter.}

\item{method}{A method to select metabolites across al batches.}

\item{assay}{An assay name to measure the missingness of the signals.}
}
\description{
Filter and order metabolites across all batches
}
