% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign.R
\name{expDesign}
\alias{expDesign}
\title{Generate experiment design}
\usage{
expDesign()
}
\arguments{
\item{n}{number of samples for the design}

\item{sr}{number of biological samples for short replicates per row}

\item{br}{number of biological samples for batch replicates}

\item{lr}{number of biological samples for long replicates}

\item{fsr}{frequency of short replicates. Default to every row}

\item{flr}{frequency of long replicates. Default to every 50 row}

\item{ncol}{number of columns for matrix}

\item{nrow}{number of samples per plate}

\item{type}{type selection of custom plate size or 96 well plate (default)}
}
\value{

}
\description{
Generate experiment design
}
