% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv.R
\name{hierarchy}
\alias{hierarchy}
\title{Hierarchical RUV}
\usage{
hierarchy(
  dat_list,
  assay,
  k = 5,
  rep,
  newAssay = NULL,
  balanced = FALSE,
  negCtl = NULL,
  hOrder = NULL
)
}
\arguments{
\item{dat_list}{A list of SummarizedExperiment data where each object
represents a batch.}

\item{assay}{An assay name to measure the missingness of the signals.}

\item{k}{Parameter \code{k} for RUV-III.}

\item{rep}{A name of the variable in the colData of each SummarizedExperiment
 data in the \code{dat_list}. This variable in colData
should be a logical vector of whic samples are a intra-batch replicate}

\item{newAssay}{A name of the new assay for cleaned (preprocessed) data.}

\item{balanced}{A type of hierarchical approach, "concatenate" or "balanced".
Default is "concatenate" approach.}

\item{negCtl}{A vector of row IDs for selection of stable matbolites to be
used as negative control in RUV.}

\item{hOrder}{A vector of batch names from names of list \code{dat_list}.}
}
\value{
A normalised data as SummarizedExperiment object.
}
\description{
Hierarchical RUV
}
